api 수정

+++ real_estate/item/models.py
@@ -6,7 +6,7 @@ class Item(models.Model):
     title = models.CharField(max_length=100)
     description = models.TextField()
     created = models.DateTimeField(auto_now_add=True)
-    price = models.IntegerField(default=0)
+    price = models.DecimalField(default=0, max_digits=12, decimal_places=8)
     image = models.ImageField(upload_to='upload/item_images/')
     owner = models.ForeignKey(User, related_name='items', on_delete=models.CASCADE, null=True)
